[NC-37314] adjust ha template to accept dev vhd, and adjusting azure-dev-deploy to use inboundHa.json
Previously, although inboundHA.json has the blob parameter, it doesn't use it when creating the VMs.
Adding two additional parameters for azure-dev-deploy: isHa, and resource_group_prefix.
If provide the 3rd param as 'ha', it will use inboundHa.json, and mainTemplate for all other values.
The 4th param will add a prefix to the resouce group, storage account and dns names. If not provided, 'dev' will be used.